66ba87e7dec2544dcecbb722e931cd19

+++ GIT_PUSH_DB.sh
@@ -2,9 +2,14 @@
 
 echo "----- Start -----"
 
-rm $PWD/dist/*.map
+
+
+if [ -e $PWD/dist/*map ]; then
+    rm $PWD/dist/*.map
+fi
+
 cp $PWD/db.json $PWD/dist/db.json
-# find $PWD/* -empty | xargs rm
+
 git pull --ff-only
 git add .
 git commit -m "$(date | md5 )"
